Add kibana3 support
Add support for the javascript version of kibana.

By default, this is done by serving both the kibana and also an
elasticsearch proxy from the same location. Allowed GETs and POSTs
for read-only access to elasticsearch are passed to a proxy and all
other requests are served from the kibana source directory.

An optional prefix, such as 'elasticsearch/', can be specified in which
case the reverse proxy to elasticsearch will be served from this
sub-path.

Change-Id: I13f9dff0bbd6498a36dc75b026c9042a9bb05e8f