Vendor dashLoader.js with Share patch
In order to make Share work without relying on elasticsearch write perms
we patch dashLoader.js to construct the urls based on the logstash
dashboard since it is deterministic anyways. This is a massive hack but
maybe it will be less terrible once ES is upgraded and we can use
kibana4.

Change-Id: I64cc693e5e188c68e134b4769ccc519f6191f36c