Configure apache when we use it
The kibana apache vhost file makes use of mod rewrite and mod proxy but
we don't enable those modules explicitly here. We happen to enable them
in other places that use this puppet module but that leads to races and
isn't explicit enough. Go ahead and explicitly enable the deps in this
puppet module so that it is self sufficient.

Change-Id: Ic05f26d61f2ac4b5adffee98782d1a02931d5731