Set server admin var so that vhost works
Using the puppet 4 future parser it is far more particular about
variables and their scoping. We don't set a serveradmin in this class
but then attempt to use it in the vhost template. This was fine under
puppet 3 as it would just look for this out of the current scope but
breaks under puppet 4.

Make this trivially happy on puppet 4 by setting a default that is sane.

Change-Id: Id415056cdfa693fd199e0b4a0c6b09e20faf23ef